Fix memory bandwidth decoding error (can be a Double)

+++ DarkbloomDashboard/Logic/API Codables/DarkbloomProviderStat.swift
@@ -12,7 +12,7 @@ struct DarkbloomProviderStat: Decodable, Equatable {
     let gpuCores: Int
     
     let memoryGb: Int
-    let memoryBandwidthGbs: Int
+    let memoryBandwidthGbs: Double
     
     let requestsServed: Int
     let tokensGenerated: Int
